Extracted text_block in its own file
